changing name from BasicMessage to Message
