+++ src/wf_core/message_tests.rs
@@ -1,9 +1,4 @@
-use crate::{
-    wf_field::definitions,
-    wf_validation::{Validation, ValidationError},
-};
-
-use super::{basic_message::BasicMessage, decode, encode};
+use super::{decode, encode};
 
 #[test]
 fn encode_sign_signal_message() {
