+++ src/wf_json/mod.rs
@@ -6,12 +6,12 @@ mod serialize;
 
 pub use deserialize::WhiteflagFieldValues;
 
-use crate::{error::WhiteflagError, wf_core::basic_message::BasicMessage};
+use crate::{error::WhiteflagError, wf_core::message::Message};
 
-impl BasicMessage {
+impl Message {
     pub fn deserialize_from_json<T: AsRef<str>>(json: T) -> Result<Self, WhiteflagError> {
         let message: WhiteflagFieldValues =
             serde_json::from_str(json.as_ref()).map_err(|e| WhiteflagError::Serde(e))?;
-        Ok(BasicMessage::compile(message.fields.as_ref()))
+        Ok(Message::compile(message.fields.as_ref()))
     }
 }
